refactor: move methods to IndexedImage
